requirements.txt
streamlit==1.12.0
pandas
numpy
plotly
psycopg2-binary
python-dotenv
torch --index-url https://download.pytorch.org/whl/cpu
transformers
huggingface_hub
